draft of encoder fp16 support

+++ sockeye/utils.py
@@ -129,14 +129,14 @@ def save_graph(symbol: mx.sym.Symbol, filename: str, hide_weights: bool = True):
     dot.render(filename=filename)
 
 
-def compute_lengths(sequence_data: mx.sym.Symbol) -> mx.sym.Symbol:
+def compute_lengths(sequence_data: mx.sym.Symbol, dtype='float32') -> mx.sym.Symbol:
     """
     Computes sequence lenghts of PAD_ID-padded data in sequence_data.
 
     :param sequence_data: Input data. Shape: (batch_size, seq_len).
     :return: Length data. Shape: (batch_size,).
     """
-    return mx.sym.sum(mx.sym.broadcast_not_equal(sequence_data, mx.sym.zeros((1,))), axis=1)
+    return mx.sym.sum(mx.sym.broadcast_not_equal(sequence_data, mx.sym.zeros((1,), dtype=dtype)), axis=1)
 
 
 def save_params(arg_params: Mapping[str, mx.nd.NDArray], fname: str,
